Updated lc.Rmd
Added the retention time model for now. We may want to replace this with a pH estimate, but for now it has the colored balls.
To do are still the peak shape evaluation, PDMS and column temperature. So this may be a couple of numbers (pH and temperature), a figure with an average peak, indicating peak shape parameters, and some barplot of PDMS signal relative to clean reference.