Added an animation for Ben's talk
I can easily modify the size of the animation to fit your PowerPoint template. We could also consider making the lines bolder and text larger. R gives pretty much complete control. To be on the safe side, we could bring an MPEG4 video as well, and test it in the speaker ready room.